+++ Changelog.md
@@ -5,8 +5,9 @@ __New Features__
 - Adds inputs for modeling skylight curbs and/or shafts.
 - Allows modeling exterior horizontal insulation for a slab-on-grade foundation (or basement/crawlspace floor).
 - Allows alternative infiltration input `AirInfiltrationMeasurement/LeakinessDescription`, in which the infiltration level is estimated using age of home, climate zone, foundation type, etc.
-- Allows optional `Window/ExteriorShading/Type` input as an alternative to `ExteriorShading/SummerShadingCoefficient` and `ExteriorShading/WinterShadingCoefficient`.
-- Allows optional `Window/InsectScreen` inputs.
+- Window shading enhancements:
+  - Allows optional `ExteriorShading/Type` input as an alternative to `ExteriorShading/SummerShadingCoefficient` and `ExteriorShading/WinterShadingCoefficient`.
+  - Allows optional `InsectScreen` input.
 - Adds optional `BuildingConstruction/UnitHeightAboveGrade` input for, e.g., apartment units of upper levels where the wind speed, and thus infiltration rate, is higher.
 - Updates operational calculations (i.e., when `NumberofResidents` provided):
   - Updates hot water usage based on FSEC study.